Move generator to oasis package
